[test] HTTP Parser 테스트 수정
HTTP Parser 의 퍼블릭 메서드 시그니처 변경에 따라 테스트 코드를 수정하였습니다.
그리고 메서드를 별도의 클래스로 분리한 부분에 있어서 해당 테스트 클래스를 따로 작성하였습니다.
